beamer: move deploy.artifacts one level up
