+++ beamer/artifacts.py
