+++ beamer/contracts.py
@@ -7,7 +7,7 @@ import web3
 from web3 import Web3
 from web3.contract import Contract
 
-import beamer.deploy.artifacts
+import beamer.artifacts
 from beamer.typing import BlockNumber, ChainId, ChecksumAddress
 
 
@@ -35,7 +35,7 @@ DeploymentInfo = dict[ChainId, dict[str, ContractInfo]]
 
 
 def prepare_deployment_infos(
-    abi_dir: Path, contracts: dict[str, beamer.deploy.artifacts.DeployedContractInfo]
+    abi_dir: Path, contracts: dict[str, beamer.artifacts.DeployedContractInfo]
 ) -> dict[str, ContractInfo]:
     abis = {}
     infos = {}
@@ -54,7 +54,7 @@ def prepare_deployment_infos(
 def load_deployment_info(artifacts_dir: Path, abi_dir: Path) -> DeploymentInfo:
     deployment_info = {}
     for artifact_path in artifacts_dir.glob("*.deployment.json"):
-        deployment = beamer.deploy.artifacts.Deployment.from_file(artifact_path)
+        deployment = beamer.artifacts.Deployment.from_file(artifact_path)
         if deployment.chain is None:
             continue
         deployment_info[deployment.chain.chain_id] = prepare_deployment_infos(
@@ -70,7 +70,7 @@ def contracts_for_web3(web3: Web3, artifacts_dir: Path, abi_dir: Path) -> dict[s
 
 
 def obtain_contract(
-    w3: Web3, abi_dir: Path, deployment: beamer.deploy.artifacts.Deployment, name: str
+    w3: Web3, abi_dir: Path, deployment: beamer.artifacts.Deployment, name: str
 ) -> Contract:
     chain_id = w3.eth.chain_id
