+++ beamer/deploy/util.py
@@ -10,7 +10,7 @@ from web3.contract import Contract, ContractConstructor
 from web3.contract.contract import ContractFunction
 
 import beamer.deploy.config as config
-from beamer.deploy.artifact import ChainDeployment, DeployedContractInfo, Deployment
+from beamer.artifacts import ChainDeployment, DeployedContractInfo, Deployment
 from beamer.typing import BlockNumber, ChainId
 from beamer.util import get_commit_id, transact
